Update C# port of Windows API
